requirements.txt
# Core dependencies
streamlit
pandas
matplotlib
seaborn
pillow
fpdf

# LangChain & LangGraph
langchain
langchain-openai
langchain-community
langgraph

# OpenAI
openai

# Environment
python-dotenv
